+++ libs/react/src/components/molecules/journeyOverviewCard/journeyOverviewCard.stories.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { withKnobs } from '@storybook/addon-knobs';
 import { Meta, StoryObj } from '@storybook/react';
 import { JourneyOverviewCard } from './journeyOverviewCard';
-import { getCategoriesHandler, StoryMockProvider } from '@coldpbc/mocks';
-import { FormSection } from '@coldpbc/components';
+import { getAssessmentsHandler, StoryMockProvider} from '@coldpbc/mocks';
+import { ColdAssessmentsProvider } from "@coldpbc/providers";
 
 const meta: Meta<typeof JourneyOverviewCard> = {
   title: 'Molecules/JourneyOverviewCard',
@@ -18,10 +18,26 @@ type Story = StoryObj<typeof meta>;
 export const Default: Story = {
   render: (args) => {
     return (
-      <StoryMockProvider handlers={[]}>
-        <div className="w-[668px]">
-          <JourneyOverviewCard />
-        </div>
+      <StoryMockProvider handlers={getAssessmentsHandler.default}>
+        <ColdAssessmentsProvider>
+          <div className="w-[668px]">
+            <JourneyOverviewCard />
+          </div>
+        </ColdAssessmentsProvider>
+      </StoryMockProvider>
+    );
+  },
+};
+
+export const SingleCompliance: Story = {
+  render: (args) => {
+    return (
+      <StoryMockProvider handlers={getAssessmentsHandler.single}>
+        <ColdAssessmentsProvider>
+          <div className="w-[668px]">
+            <JourneyOverviewCard />
+          </div>
+        </ColdAssessmentsProvider>
       </StoryMockProvider>
     );
   },
@@ -29,10 +45,12 @@ export const Default: Story = {
 
 export const EmptyData = () => {
   return (
-    <StoryMockProvider handlers={[getCategoriesHandler.empty]}>
-      <div className="w-[668px]">
-        <JourneyOverviewCard />
-      </div>
+    <StoryMockProvider handlers={getAssessmentsHandler.empty}>
+      <ColdAssessmentsProvider>
+        <div className="w-[668px]">
+          <JourneyOverviewCard />
+        </div>
+      </ColdAssessmentsProvider>
     </StoryMockProvider>
   );
 };