+++ libs/react/src/components/pages/subcategoryActionsList/subcategoryActionsList.stories.tsx
@@ -66,22 +66,6 @@ export const Default: Story = {
       await expect(text2).toBeInTheDocument();
     });
 
-    await step('Category Score', async () => {
-      // find Category Score text
-      const text = canvas.getByText('Facilities Score');
-      await expect(text).toBeInTheDocument();
-      // find Category Score number
-      const text2 = canvas.getByText('25%');
-      await expect(text2).toBeInTheDocument();
-    });
-
-    await step('Footprint Chart', async () => {
-      // find Action Score text
-      const tCO2eScores = await canvas.findAllByText('tCO2e');
-      // check if the tc02e scores array are not empty
-      await expect(tCO2eScores).not.toHaveLength(0);
-    });
-
     await step('Check Subcategory Description', async () => {
       const text = canvas.getByTestId('subcategory-description');
       await expect(text).toBeInTheDocument();