+++ CHANGELOG.md
@@ -1,29 +1,112 @@
-# [1.157.0](https://github.com/ColdPBC/cold-monorepo/compare/v1.156.0...v1.157.0) (2024-03-11)
-
-
-### Features
-
-* have subtract icon when a question is complete for a category ([12b97a4](https://github.com/ColdPBC/cold-monorepo/commit/12b97a4df5c0a6ed408f470dee3df3d3998b9428))
-
-# [1.156.0](https://github.com/ColdPBC/cold-monorepo/compare/v1.155.0...v1.156.0) (2024-03-11)
-
-
-### Features
-
-* add rounded to survey compliance flow step ([8cd374f](https://github.com/ColdPBC/cold-monorepo/commit/8cd374fe338f45c7b67351ed3545b769d1a62309))
-
-# [1.155.0](https://github.com/ColdPBC/cold-monorepo/compare/v1.154.0...v1.155.0) (2024-03-11)
+## [1.162.1](https://github.com/ColdPBC/cold-monorepo/compare/v1.162.0...v1.162.1) (2024-03-13)
 
 
 ### Bug Fixes
 
-* make filter code a bit cleaner ([bb815b2](https://github.com/ColdPBC/cold-monorepo/commit/bb815b2971c19d8fb0e719df878dce7d631045d6))
-* tweak prompts ([b5768d0](https://github.com/ColdPBC/cold-monorepo/commit/b5768d061b0ce08de623c73e0f6469280928d4ce))
+* bug in policies throwing 500 error ([af83962](https://github.com/ColdPBC/cold-monorepo/commit/af839624dbc2a359ba880b3d202a8a24dab64c56))
+
+# [1.162.0](https://github.com/ColdPBC/cold-monorepo/compare/v1.161.1...v1.162.0) (2024-03-13)
 
 
 ### Features
 
-* support multiple file uploads ([16c417b](https://github.com/ColdPBC/cold-monorepo/commit/16c417b00a4a40f846dfd10081453d933283819e))
+* resolve issue with list item input when values already provided ([58bba97](https://github.com/ColdPBC/cold-monorepo/commit/58bba97f59cc904cd9b9fd5937ed35e9897203a0))
+* update list item input design ([db19d00](https://github.com/ColdPBC/cold-monorepo/commit/db19d00881a2b8dfd8f4487236a340cd795353b5))
+
+## [1.161.1](https://github.com/ColdPBC/cold-monorepo/compare/v1.161.0...v1.161.1) (2024-03-13)
+
+
+### Bug Fixes
+
+* allow select when value provided is a string ([b3675fa](https://github.com/ColdPBC/cold-monorepo/commit/b3675fab8e63ce8b596f0a0daca7c5970746a337))
+* dont show selected options when the value is not an array ([93faf3d](https://github.com/ColdPBC/cold-monorepo/commit/93faf3df9f27ba969e69fd102cd43b129ae7d691))
+
+## [1.161.1-COLD-621.1](https://github.com/ColdPBC/cold-monorepo/compare/v1.161.0...v1.161.1-COLD-621.1) (2024-03-13)
+
+
+### Bug Fixes
+
+* allow select when value provided is a string ([b3675fa](https://github.com/ColdPBC/cold-monorepo/commit/b3675fab8e63ce8b596f0a0daca7c5970746a337))
+
+# [1.161.0](https://github.com/ColdPBC/cold-monorepo/compare/v1.160.0...v1.161.0) (2024-03-13)
+
+
+### Bug Fixes
+
+* remove broken test ([c5efc76](https://github.com/ColdPBC/cold-monorepo/commit/c5efc765cfb2b0d2ef77cb3089e2d0052e3cfeb9))
+* Remove journey module on subcategory actions list page ([f4cf554](https://github.com/ColdPBC/cold-monorepo/commit/f4cf554ec0ab87abb29cde6d1262aa84486ab8b6))
+
+
+### Features
+
+* Update AI message look and feel and add it for additional context ([5dd8798](https://github.com/ColdPBC/cold-monorepo/commit/5dd8798fa8da2657556dc8c6fc6ab7997cf71db6))
+
+# [1.160.0](https://github.com/ColdPBC/cold-monorepo/compare/v1.159.1...v1.160.0) (2024-03-13)
+
+
+### Features
+
+* show ai icon if question is ai answered ([fd2a17e](https://github.com/ColdPBC/cold-monorepo/commit/fd2a17e9cb53b90a7ea076b7c92c9fe589888b0a))
+
+## [1.159.1](https://github.com/ColdPBC/cold-monorepo/compare/v1.159.0...v1.159.1) (2024-03-12)
+
+
+### Bug Fixes
+
+* bug that deleted survey rather than survey data ([b3b8734](https://github.com/ColdPBC/cold-monorepo/commit/b3b87346ed91bbb0955138023789a4fedd8325d1))
+
+# [1.159.0](https://github.com/ColdPBC/cold-monorepo/compare/v1.158.0...v1.159.0) (2024-03-12)
+
+
+### Bug Fixes
+
+* Fix test related to Assessments ([7b12ec7](https://github.com/ColdPBC/cold-monorepo/commit/7b12ec78cdb88df9f37ff619f3810ae085280c51))
+* properly support multiple urls in OrgSWR ([6b8cc3a](https://github.com/ColdPBC/cold-monorepo/commit/6b8cc3a790d380444452632165291069f10439c2))
+
+
+### Features
+
+* Finish adjustment of Journey component to Assessment component and add stories ([0827305](https://github.com/ColdPBC/cold-monorepo/commit/0827305471e1d98acdf1042ac10e5790df91ebaf))
+* set up assessments page to use compliance info ([1641b5b](https://github.com/ColdPBC/cold-monorepo/commit/1641b5b14844c32902dc7b5add613f43af67a8ce))
+
+# [1.158.0](https://github.com/ColdPBC/cold-monorepo/compare/v1.157.0...v1.158.0) (2024-03-12)
+
+
+### Features
+
+* mqtt updates ([59acc3b](https://github.com/ColdPBC/cold-monorepo/commit/59acc3ba771f2d45c63c29939181c7cb221fbb3f))
+
+# [1.158.0-COLD-573.1](https://github.com/ColdPBC/cold-monorepo/compare/v1.157.0...v1.158.0-COLD-573.1) (2024-03-11)
+
+
+### Features
+
+* mqtt updates ([59acc3b](https://github.com/ColdPBC/cold-monorepo/commit/59acc3ba771f2d45c63c29939181c7cb221fbb3f))
+
+# [1.157.0](https://github.com/ColdPBC/cold-monorepo/compare/v1.156.0...v1.157.0) (2024-03-11)
+
+
+### Bug Fixes
+
+* properly support multiple urls in OrgSWR ([6b8cc3a](https://github.com/ColdPBC/cold-monorepo/commit/6b8cc3a790d380444452632165291069f10439c2))
+
+
+### Features
+
+* Finish adjustment of Journey component to Assessment component and add stories ([0827305](https://github.com/ColdPBC/cold-monorepo/commit/0827305471e1d98acdf1042ac10e5790df91ebaf))
+* set up assessments page to use compliance info ([1641b5b](https://github.com/ColdPBC/cold-monorepo/commit/1641b5b14844c32902dc7b5add613f43af67a8ce))
+
+# [1.155.0-cold-522.1](https://github.com/ColdPBC/cold-monorepo/compare/v1.154.0...v1.155.0-cold-522.1) (2024-03-08)
+
+
+### Bug Fixes
+
+* properly support multiple urls in OrgSWR ([6b8cc3a](https://github.com/ColdPBC/cold-monorepo/commit/6b8cc3a790d380444452632165291069f10439c2))
+
+
+### Features
+
+* set up assessments page to use compliance info ([1641b5b](https://github.com/ColdPBC/cold-monorepo/commit/1641b5b14844c32902dc7b5add613f43af67a8ce))
 
 # [1.154.0](https://github.com/ColdPBC/cold-monorepo/compare/v1.153.0...v1.154.0) (2024-03-07)